leaf-health-check/requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.1.78
numpy>=1.24.3
pillow>=10.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
